requirements.txt
streamlit
google-generativeai
youtube-transcript-api==1.2.2
langchain
langchain-google-genai
langchain-chroma
python-dotenv
google-genai
gtts
langchain
langchain_community
langchain-google-genai
python-dotenv
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
rapidocr-onnxruntime
pypdf
pysqlite3-binary
langchain
langchain-huggingface
sentence-transformers
torch
requests

HuggingFaceEmbeddings